GFC-?: Chipped away at ValidationService. Produced 2 new classes from it
GFC-?: Broke down ComponentsValidation. Still chipping

GFC-?: Extracted methods from AddressValidation. Cleaned up code.
