+++ test/uk/gov/hmrc/gform/services/AddressValidationSpec.scala
@@ -21,7 +21,8 @@ import cats.scalatest.EitherMatchers
 import cats.scalatest.ValidatedValues._
 import org.scalatest.concurrent.ScalaFutures
 import org.scalatest.mockito.MockitoSugar.mock
-import org.scalatest.{ FlatSpec, Matchers }
+import org.scalatest.{FlatSpec, Matchers}
+
 import scala.concurrent.ExecutionContext.Implicits.global
 import uk.gov.hmrc.gform.GraphSpec
 import uk.gov.hmrc.gform.auth.models.MaterialisedRetrievals
@@ -30,8 +31,8 @@ import uk.gov.hmrc.gform.sharedmodel.ExampleData
 import uk.gov.hmrc.gform.sharedmodel.form.ThirdPartyData
 import uk.gov.hmrc.gform.validation.ValidationUtil.ValidatedType
 import uk.gov.hmrc.gform.sharedmodel.form.EnvelopeId
-import uk.gov.hmrc.gform.sharedmodel.formtemplate.{ Address, FormComponent, FormComponentId }
-import uk.gov.hmrc.gform.validation.{ ComponentsValidator, ValidationValues }
+import uk.gov.hmrc.gform.sharedmodel.formtemplate.{Address, FormComponent, FormComponentId}
+import uk.gov.hmrc.gform.validation.{ComponentsValidator, ValidationValues}
 import uk.gov.hmrc.http.HeaderCarrier
 
 class AddressValidationSpec extends FlatSpec with Matchers with EitherMatchers with ScalaFutures with GraphSpec {
